requirements.txt
numpy
pandas
scikit-learn
scipy

torch
tqdm
rich

matplotlib
seaborn

pyyaml
joblib
python-dotenv
requests

# notebooks  
jupyter
ipykernel
ipywidgets

# offline LLM
ollama

# LLM Report
python-docx